added user version of getUserFieldsFn
