+++ waspc/data/Generator/templates/server/src/routes/auth/passport/google/googleDefaults.js
@@ -1,4 +1,4 @@
-import { generateAvailableDictionaryUsername } from '../../utils.js'
+import { generateAvailableDictionaryUsername } from '../../../../core/auth.js'
 
 // Default implementation if there is no `auth.methods.google.configFn`.
 export function configFn() {
@@ -17,7 +17,7 @@ export function configFn() {
 }
 
 // Default implementation if there is no `auth.methods.google.getUserFields`.
-export async function getUserFields(_context, args) {
+export async function getUserFields(_context, _args) {
   const username = await generateAvailableDictionaryUsername()
   // const username = await generateAvailableUsername(args.profile.displayName.split(' '), { separator: '.' })
   // TODO: rename `email` to `username` after merge.
