+++ waspc/data/Generator/templates/server/src/routes/auth/utils.js
@@ -1,7 +1,6 @@
 {{={= =}=}}
 
 import { v4 as uuidv4 } from 'uuid'
-import { randomInt } from 'node:crypto'
 
 import prisma from '../../dbClient.js'
 
@@ -39,48 +38,4 @@ export async function findOrCreateUserBySocialLogin(provider, providerId, getUse
   }
 
   return await prisma.{= userEntityLower =}.create({ data: userAndSocialLogin })
-}
-
-export async function generateAvailableDictionaryUsername() {
-  const adjectives = ['fuzzy', 'tall', 'short', 'nice', 'happy', 'quick', 'slow', 'good', 'new', 'old', 'first', 'last', 'old', 'young']
-  const colors = ['red', 'green', 'blue', 'white', 'black', 'brown', 'purple', 'orange', 'yellow']
-  const nouns = ['cat', 'dog', 'lion', 'rabbit', 'duck', 'pig', 'bee', 'goat', 'crab', 'fish', 'chicken', 'horse', 'llama', 'camel', 'sheep']
-
-  const potentialUsernames = []
-  for (let i = 0; i < 10; i++) {
-    const potentialUsername = `${adjectives[randomInt(adjectives.length)]}-${colors[randomInt(colors.length)]}-${nouns[randomInt(nouns.length)]}-${randomInt(100_000)}`
-    potentialUsernames.push(potentialUsername)
-  }
-
-  return findAvailableUsername(potentialUsernames)
-}
-
-export async function generateAvailableUsername(strings, config) {
-  const separator = config?.separator || '-'
-  const baseUsername = strings.join(separator)
-
-  const potentialUsernames = []
-  for (let i = 0; i < 10; i++) {
-    const potentialUsername = `${baseUsername}${separator}${randomInt(100_000)}`
-    potentialUsernames.push(potentialUsername)
-  }
-
-  return findAvailableUsername(potentialUsernames)
-}
-
-async function findAvailableUsername(potentialUsernames) {
-  // TODO: Change `email` to `username` after merge.
-  const users = await prisma.{= userEntityLower =}.findMany({
-    where: {
-      email: { in: potentialUsernames },
-    }
-  })
-  const takenUsernames = users.map(user => user.email)
-  const availableUsernames = potentialUsernames.filter(username => !takenUsernames.includes(username))
-
-  if (availableUsernames.length === 0) {
-    throw 'Unable to generate a unique username. Please contact Wasp.'
-  }
-
-  return availableUsernames[0]
 }
